# OPENAPI.yaml
openapi: 3.0.3
info:
  title: HMR Backend API
  description: API for managing real estate tokenization, organizations, users, investments, and wallet transactions.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /organizations:
    get:
      tags:
        - Organizations
      summary: List all organizations
      responses:
        '200':
          description: A list of all organizations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
    post:
      tags:
        - Organizations
      summary: Create a new organization
      description: Creates a new organization with an auto-generated displayCode and initial liquidity of 0.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreate'
      responses:
        '201':
          description: The organization was successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
  /organizations/{id}:
    get:
      tags:
        - Organizations
      summary: Get organization by ID
      description: Get organization by its UUID or displayCode (e.g., ORG-000001).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The UUID or displayCode of the organization.
      responses:
        '200':
          description: Organization details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          description: Organization not found.
  /organizations/{id}/liquidity:
    get:
      tags:
        - Organizations
      summary: Get organization liquidity analytics
      description: Get liquidity analytics for a specific organization by its UUID or displayCode.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The UUID or displayCode of the organization.
      responses:
        '200':
          description: Organization liquidity details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationLiquidity'
  /organizations/{id}/transactions:
    get:
      tags:
        - Organizations
      summary: Get all transactions for an organization
      description: Get all transactions for an organization by its UUID or displayCode.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The UUID or displayCode of the organization.
      responses:
        '200':
          description: A list of transactions for the organization.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'

  # ----------------------------------------------
  # Users (Admin)
  # ----------------------------------------------
...
  /org/auth/login:
    post:
      tags: [Organization Admin Auth]
      summary: Organization admin login (no JWT returned)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Login result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgAdminLoginResponse'
  /org/auth/change-password/{adminId}:
    patch:
      tags: [Organization Admin Auth]
      summary: Change organization admin password
      parameters:
        - name: adminId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword: { type: string }
                newPassword: { type: string }
      responses:
        '200':
          description: Change password result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }

  # ----------------------------------------------
  # Users (Direct Management)
  # ----------------------------------------------
...
  /investments/analytics/organization/{orgId}:
    get:
      tags:
        - Investments
      summary: Get organization investment analytics
      description: Get comprehensive investment analytics for a specific organization including all investments in their properties and sum analytics.
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
          description: The UUID or displayCode of the organization (e.g., ORG-000001).
      responses:
        '200':
          description: Organization investment analytics with detailed breakdown.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationInvestmentAnalytics'
        '404':
          description: Organization not found.
  /investments/analytics/user/{userId}/organization/{orgId}:
    get:
      tags:
        - Investments
      summary: Get user-organization investment analytics
      description: Get comprehensive investment analytics for a specific user within a specific organization including detailed investment list and sum analytics.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: The UUID or displayCode of the user (e.g., USR-000001).
        - name: orgId
          in: path
          required: true
          schema:
            type: string
          description: The UUID or displayCode of the organization (e.g., ORG-000001).
      responses:
        '200':
          description: User-organization investment analytics with detailed breakdown.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOrganizationInvestmentAnalytics'
        '404':
          description: User or organization not found.

  # ----------------------------------------------
  # Portfolio
  # ----------------------------------------------
...
components:
  schemas:
...
    OrgAdminLoginResponse:
      type: object
      properties:
        organizationId: { type: string }
        admin:
          type: object
          properties:
            id: { type: string }
            email: { type: string }
            fullName: { type: string, nullable: true }
    # ----------------------------------------------
    # ENTITIES
    # ----------------------------------------------
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        displayCode:
          type: string
          example: ORG-000001
        name:
          type: string
          example: HMR Builders
        description:
          type: string
          example: Leading real estate developer
        website:
          type: string
          format: uri
          example: https://hmrbuilders.com
        logoUrl:
          type: string
          format: uri
          example: https://example.com/logo.png
        liquidityUSDT:
          type: string
          format: decimal
          description: Total liquidity held by the organization.
          example: '250000.000000'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Investment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        displayCode:
          type: string
          example: INV-000001
        userId:
          type: string
          format: uuid
        propertyId:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        amountUSDT:
          type: string
          format: decimal
          example: '2500.000000'
        tokensOwned:
          type: string
          format: decimal
          example: '2.500000'
        status:
          type: string
          enum: [confirmed, pending]
          example: confirmed
        createdAt:
          type: string
          format: date-time
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        displayCode:
          type: string
          example: TXN-000005
        type:
          type: string
          enum: [inflow, outflow, investment, reward, deposit, withdrawal]
          example: inflow
        fromEntity:
          type: string
          example: Ali Khan
        toEntity:
          type: string
          example: HMR Builders
        amountUSDT:
          type: string
          format: decimal
          example: '2500.000000'
        propertyId:
          type: string
          format: uuid
          nullable: true
        organizationId:
          type: string
          format: uuid
          nullable: true
        userId:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [completed, pending, failed]
          example: completed
        description:
          type: string
          example: Liquidity inflow from Ali Khan
        createdAt:
          type: string
          format: date-time
...
    OrganizationCreate:
      type: object
      properties:
        name:
          type: string
          example: HMR Builders
        description:
          type: string
          example: Leading real estate developer
        website:
          type: string
          format: uri
          example: https://hmrbuilders.com
        logoUrl:
          type: string
          format: uri
          example: https://example.com/logo.png
      required:
        - name
...
    OrganizationLiquidity:
      type: object
      properties:
        organizationId:
          type: string
          example: ORG-000001
        organizationName:
          type: string
          example: HMR Builders
        liquidityUSDT:
          type: string
          format: decimal
          example: '250000.000000'
        lastUpdated:
          type: string
          format: date-time

    # ----------------------------------------------
    # DASHBOARD & ANALYTICS SCHEMAS
    # ----------------------------------------------
...
    InvestmentAnalytics:
      type: object
      properties:
        totalInvestments:
          type: integer
          description: Total number of investments
          example: 5
        totalAmountUSDT:
          type: string
          format: decimal
          description: Sum of all investment amounts in USDT
          example: "12500.000000"
        totalTokensPurchased:
          type: string
          format: decimal
          description: Sum of all tokens purchased
          example: "12.500000"
        averageInvestmentAmount:
          type: string
          format: decimal
          description: Average investment amount per investment
          example: "2500.000000"
        averageTokensPerInvestment:
          type: string
          format: decimal
          description: Average tokens per investment
          example: "2.500000"
        totalExpectedROI:
          type: string
          format: decimal
          description: Sum of all expected ROI percentages
          example: "50.00"
        activeInvestments:
          type: integer
          description: Number of investments with active status
          example: 3
        completedInvestments:
          type: integer
          description: Number of investments with confirmed status
          example: 2
        pendingInvestments:
          type: integer
          description: Number of investments with pending status
          example: 0
        totalValueAtCurrentPrice:
          type: string
          format: decimal
          description: Current value of all tokens at current property prices
          example: "12500.000000"
...
    OrganizationInvestmentAnalytics:
      type: object
      properties:
        organization:
          type: object
          properties:
            id:
              type: string
              format: uuid
            displayCode:
              type: string
              example: ORG-000001
            name:
              type: string
              example: HMR Builders
        investments:
          type: array
          items:
            $ref: '#/components/schemas/Investment'
        analytics:
          $ref: '#/components/schemas/InvestmentAnalytics'
    UserOrganizationInvestmentAnalytics:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            displayCode:
              type: string
              example: USR-000001
            fullName:
              type: string
              example: John Doe
            email:
              type: string
              format: email
              example: john@example.com
        organization:
          type: object
          properties:
            id:
              type: string
              format: uuid
            displayCode:
              type: string
              example: ORG-000001
            name:
              type: string
              example: HMR Builders
        investments:
          type: array
          items:
            $ref: '#/components/schemas/Investment'
        analytics:
          $ref: '#/components/schemas/InvestmentAnalytics'